+++ apps/api/services/branch_service.py
@@ -1,41 +1,51 @@
-from apps.core.models import Branch
+from django.db.models import QuerySet
+
+from apps.core.models import Branch, User
 
 from .company_services import verification_owner_company_service
 
 
-def get_branch_by_id_service(user: object, branch_id: int) -> object | None:
+def get_branch_by_id_service(user: QuerySet[User], branch_id: int) -> QuerySet[Branch] | None:
     """
     Функция возвращает филиал по его id. Если текущий пользователь не владелец этого филиала, вернется None
     :param user: Текущий пользователь
     :param branch_id: id филиала
     :return: optional: branch
     """
-    branch: object = Branch.objects.get(pk=branch_id)
-    result: object | None = (None, branch)[verification_owner_company_service(company_id=branch.company.id, user=user)]
+    branch: QuerySet[Branch] = Branch.objects.get(pk=branch_id)
+    result: QuerySet[Branch] | None = (
+        None,
+        branch
+    )[
+        verification_owner_company_service(
+            company_id=branch.company.id,
+            user=user
+        )
+    ]
     return result
 
 
-def verification_owner_branch_service(user: object, branch_id: int) -> bool:
+def verification_owner_branch_service(user: QuerySet[User], branch_id: int) -> bool:
     """
     Функция проверяет, имеет ли текущий пользователь доступ к branch
     :param user: текущий пользователь
     :param branch_id: id branch
     :return: True или False
     """
-    branch: object = Branch.objects.get(pk=branch_id)
+    branch: QuerySet[Branch] = Branch.objects.get(pk=branch_id)
     return verification_owner_company_service(user=user, company_id=branch.company_id)
 
 
-def get_all_not_detected_branch_service() -> object:
+def get_all_not_detected_branch_service() -> QuerySet[Branch]:
     """
     Функция возвращает все branch, которые еще не обработаны системой.
     :return: branch_list
     """
-    branch_list: object = Branch.objects.filter(is_detected=False)
+    branch_list: QuerySet[Branch] = Branch.objects.filter(is_detected=False)
     return branch_list
 
 
-def get_branch_service(branch_id: int) -> object:
+def get_branch_service(branch_id: int) -> QuerySet[Branch]:
     """
     Функция возвращает branch по id
     :param branch_id: id branch
@@ -45,7 +55,7 @@ def get_branch_service(branch_id: int) -> object:
 
 
 def create_branch_by_company(
-        user: object,
+        user: QuerySet[User],
         company_id: int,
         name: str,
         address: str = '',
@@ -53,7 +63,7 @@ def create_branch_by_company(
         email: str = None,
         description: str = '',
         short_description: str = ''
-) -> object | None:
+) -> QuerySet[Branch] | None:
     """
     Функция создает branch, если у текущего пользователя есть доступ
     :param user: Текущий пользователь
@@ -66,7 +76,7 @@ def create_branch_by_company(
     :param short_description: короткое описание
     :return: optional: branch
     """
-    result: object | None = (
+    result: QuerySet[Branch] | None = (
         None,
         _create_branch_service(
             company_id=company_id,
@@ -89,8 +99,8 @@ def _create_branch_service(
         email: str,
         description: str,
         short_description: str
-) -> object:
-    branch: object = Branch.objects.create(
+) -> QuerySet[Branch]:
+    branch: QuerySet[Branch] = Branch.objects.create(
         company_id=company_id,
         name=name,
         address=address,