+++ apps/api/services/rate_info_service.py
@@ -1,7 +1,9 @@
+from django.db.models import QuerySet
+
 from apps.core.models import RateInfo
 
 
-def get_rate_info_service() -> object:
+def get_rate_info_service() -> QuerySet[RateInfo]:
     """
     Функция возвращает последний rate info
     :return: rate info