+++ apps/api/services/review_settings_service.py
@@ -1,9 +1,15 @@
-from apps.core.models import ReviewSettings
+from django.db.models import QuerySet
+
+from apps.core.models import ReviewSettings, User
 
 from .company_services import verification_owner_company_service
 
 
-def create_review_settings_by_company_id_service(user: object, company_id: int, mask: str) -> object | None:
+def create_review_settings_by_company_id_service(
+        user: QuerySet[User],
+        company_id: int,
+        mask: str
+) -> QuerySet[ReviewSettings] | None:
     """
     Функция создает review settings если у пользователя есть доступ, иначе вернет None
     :param user: Текущий пользователь
@@ -11,7 +17,7 @@ def create_review_settings_by_company_id_service(user: object, company_id: int, 
     :param mask: Маска
     :return: optional: review settings
     """
-    result: object | None = (
+    result: QuerySet[ReviewSettings] | None = (
         None,
         ReviewSettings.objects.create(
             company_id=company_id,
@@ -22,15 +28,15 @@ def create_review_settings_by_company_id_service(user: object, company_id: int, 
     return result
 
 
-def get_review_settings_by_id_service(user: object, review_settings_id: int) -> object | None:
+def get_review_settings_by_id_service(user: QuerySet[User], review_settings_id: int) -> QuerySet[ReviewSettings] | None:
     """
     Функция возвращает review settings если у пользователя есть доступ, иначе вернет False
     :param user: Текущий пользователь
     :param review_settings_id: review settings id
     :return: optional: review settings
     """
-    review_settings: object = ReviewSettings.objects.get(pk=review_settings_id)
-    result: object | None = (
+    review_settings: QuerySet[ReviewSettings] = ReviewSettings.objects.get(pk=review_settings_id)
+    result: QuerySet[ReviewSettings] | None = (
         None,
         review_settings
     )[verification_owner_company_service(user=user, company_id=review_settings.company_id)]